requirements.txt
pytest==8.3.2
pytest-xdist==3.6.1
pytest-order==1.3.0
pytest-dotenv==0.5.2
allure-pytest==2.13.5


# HTTP
httpx==0.27.2
pydantic==2.8.2
python-dotenv==1.0.1


# Schemathesis (property-based testing against OpenAPI)
schemathesis==4.1.4
hypothesis==6.112.1


# Pact (contract testing)
pact-python==2.3.3


# Performance
locust==2.31.6